Split util -> file_util for file manipulation
